Исправил SQL-запросы. Продолжить видео на 19:00

+++ src/main/java/com/itstep/travelpackage/model/entity/Travel.java
@@ -21,7 +21,7 @@ public class Travel extends BaseEntity {
     private Long travelFeedId;
 
     @Column(name = "days_count")
-    private Long daysCount;
+    private Integer daysCount;
 
     @ManyToOne
     @JoinColumn(
